Move actions_panel.py into RimSort.views.sub_views
